cli: split lib command shell support into child modules
